Extract monster race I/O part into the IO package
