+++ README.md
@@ -103,3 +103,4 @@
 1. SecureRandom 클래스를 제거했습니다.
 2. Menu 선택을 해야해서 Enum을 사용한 메뉴선택 방식을 사용하고자 했습니다.
 3. 메뉴 선택기능을 추가하였습니다.
+4. Input Output 을 담당하는 부분을 IO 패키지로 이동하였습니다.
