+++ src/main/java/dev/idion/monsterrace/Game.java
@@ -1,5 +1,8 @@
 package dev.idion.monsterrace;
 
+import dev.idion.monsterrace.IO.Input;
+import dev.idion.monsterrace.IO.Printer;
+
 import java.util.*;
 import java.util.stream.IntStream;
